refactor: simplify config import path
