+++ aikg/config/__init__.py
@@ -1,0 +1,3 @@
+from aikg.config.chat import ChatConfig
+from aikg.config.chroma import ChromaConfig
+from aikg.config.sparql import SparqlConfig
