+++ tests/test_load_data.py
@@ -1,9 +1,6 @@
-from aikg.config.chroma import ChromaConfig
-from aikg.config.sparql import SparqlConfig
+from aikg.config import ChromaConfig, SparqlConfig
 from aikg.flows.chroma_build import chroma_build_flow
 from aikg.flows.insert_triples import sparql_insert_flow
-
-
 
 
 def test_init_data(schema_file, small_instance_file):
